+++ ecs/archetype.go
@@ -49,7 +49,26 @@ func (a *archetypeNode) SetTransitionRemove(id ID, to *archetypeNode) {
 	a.toRemove[id] = to
 }
 
-type archetypes = pagedArr32[archetype]
+type archetypeArr = pagedArr32[archetype]
+
+type archetypes interface {
+	Get(index int) *archetype
+	Len() int
+}
+
+type singleArchetype struct {
+	archetype *archetype
+}
+
+// Get returns the value at the given index.
+func (s singleArchetype) Get(index int) *archetype {
+	return s.archetype
+}
+
+// Len returns the current number of items in the paged array.
+func (s singleArchetype) Len() int {
+	return 1
+}
 
 // Helper for accessing data from an archetype
 type archetypeAccess struct {
@@ -159,13 +178,26 @@ func (a *archetype) Init(node *archetypeNode, capacityIncrement int, forStorage 
 // Add adds an entity with zeroed components to the archetype
 func (a *archetype) Alloc(entity Entity, zero bool) uintptr {
 	idx := uintptr(a.len)
-	a.extend()
-	a.addEntity(&entity, idx)
+	a.extend(1)
+	a.addEntity(idx, &entity)
 	if zero {
 		a.ZeroAll(idx)
 	}
 	a.len++
 	return idx
+}
+
+// Add adds storage to the archetype
+func (a *archetype) AllocN(count uint32, zero bool) {
+	idx := uintptr(a.len)
+	a.extend(count)
+	if zero {
+		var i uint32
+		for i = 0; i < count; i++ {
+			a.ZeroAll(idx + uintptr(i))
+		}
+	}
+	a.len += count
 }
 
 // Add adds an entity with components to the archetype
@@ -175,8 +207,8 @@ func (a *archetype) Add(entity Entity, components ...Component) uintptr {
 	}
 	idx := uintptr(a.len)
 
-	a.extend()
-	a.addEntity(&entity, idx)
+	a.extend(1)
+	a.addEntity(idx, &entity)
 	for _, c := range components {
 		lay := a.getLayout(c.ID)
 		dst := a.Get(uintptr(idx), c.ID)
@@ -232,6 +264,11 @@ func (a *archetype) Zero(index uintptr, id ID) {
 		*(*byte)(dst) = 0
 		dst = unsafe.Add(dst, 1)
 	}
+}
+
+// SetEntity overwrites an entity
+func (a *archetype) SetEntity(index uintptr, entity Entity) {
+	a.addEntity(index, &entity)
 }
 
 // Set overwrites a component with the data behind the given pointer
@@ -321,11 +358,15 @@ func (a *archetype) copy(src, dst unsafe.Pointer, itemSize uintptr) {
 }
 
 // extend the memory buffers if necessary for adding an entity.
-func (a *archetype) extend() {
-	if a.cap > a.len {
+func (a *archetype) extend(by uint32) {
+	required := a.len + by
+	if a.cap >= required {
 		return
 	}
-	a.cap = a.capacityIncrement * ((a.cap + a.capacityIncrement) / a.capacityIncrement)
+	a.cap = a.capacityIncrement * (required / a.capacityIncrement)
+	if required%a.capacityIncrement != 0 {
+		a.cap += a.capacityIncrement
+	}
 
 	old := a.entityBuffer
 	a.entityBuffer = reflect.New(reflect.ArrayOf(int(a.cap), entityType)).Elem()
@@ -346,7 +387,7 @@ func (a *archetype) extend() {
 }
 
 // Adds an entity at the given index. Does not extend the entity buffer.
-func (a *archetype) addEntity(entity *Entity, index uintptr) {
+func (a *archetype) addEntity(index uintptr, entity *Entity) {
 	dst := unsafe.Add(a.entityPointer, entitySize*index)
 	src := reflect.ValueOf(entity).UnsafePointer()
 	a.copy(src, dst, entitySize)