+++ ecs/world_test.go
@@ -359,6 +359,147 @@ func TestWorldIter(t *testing.T) {
 	assert.Panics(t, func() { query.Close() })
 }
 
+func TestWorldNewEntities(t *testing.T) {
+	world := NewWorld(NewConfig().WithCapacityIncrement(16))
+	world.SetListener(func(e EntityEvent) {})
+
+	posID := ComponentID[position](&world)
+	rotID := ComponentID[rotation](&world)
+
+	world.NewEntity(posID, rotID)
+
+	assert.Panics(t, func() { world.newEntities(0, posID, rotID) })
+
+	query := world.newEntities(100, posID, rotID)
+	assert.Equal(t, 100, query.Count())
+
+	cnt := 0
+	for query.Next() {
+		pos := (*position)(query.Get(posID))
+		pos.X = cnt + 1
+		pos.Y = cnt + 1
+		cnt++
+	}
+	assert.Equal(t, 100, cnt)
+
+	query = world.Query(All(posID, rotID))
+	assert.Equal(t, 101, query.Count())
+
+	cnt = 0
+	for query.Next() {
+		pos := (*position)(query.Get(posID))
+		assert.Equal(t, cnt, pos.X)
+		cnt++
+	}
+
+	world.Reset()
+
+	query = world.newEntities(100, posID, rotID)
+	assert.Equal(t, 100, query.Count())
+
+	entities := make([]Entity, query.Count())
+	cnt = 0
+	for query.Next() {
+		entities[cnt] = query.Entity()
+		cnt++
+	}
+	assert.Equal(t, 100, cnt)
+
+	for _, e := range entities {
+		world.RemoveEntity(e)
+	}
+	query = world.newEntities(100, posID, rotID)
+	query.Close()
+}
+
+func TestWorldNewEntitiesWith(t *testing.T) {
+	world := NewWorld(NewConfig().WithCapacityIncrement(16))
+	world.SetListener(func(e EntityEvent) {})
+
+	posID := ComponentID[position](&world)
+	rotID := ComponentID[rotation](&world)
+
+	comps := []Component{
+		{ID: posID, Comp: &position{100, 200}},
+		{ID: rotID, Comp: &rotation{300}},
+	}
+
+	world.NewEntity(posID, rotID)
+
+	assert.Panics(t, func() { world.newEntitiesWith(0, comps...) })
+	query := world.newEntitiesWith(1)
+	query.Close()
+
+	query = world.newEntitiesWith(100, comps...)
+	assert.Equal(t, 100, query.Count())
+
+	cnt := 0
+	for query.Next() {
+		pos := (*position)(query.Get(posID))
+		assert.Equal(t, 100, pos.X)
+		assert.Equal(t, 200, pos.Y)
+		pos.X = cnt + 1
+		pos.Y = cnt + 1
+		cnt++
+	}
+	assert.Equal(t, 100, cnt)
+
+	query = world.Query(All(posID, rotID))
+	assert.Equal(t, 101, query.Count())
+
+	cnt = 0
+	for query.Next() {
+		pos := (*position)(query.Get(posID))
+		assert.Equal(t, cnt, pos.X)
+		cnt++
+	}
+
+	world.Reset()
+
+	query = world.newEntitiesWith(100,
+		Component{ID: posID, Comp: &position{100, 200}},
+		Component{ID: rotID, Comp: &rotation{300}},
+	)
+	assert.Equal(t, 100, query.Count())
+
+	cnt = 0
+	for query.Next() {
+		cnt++
+	}
+	assert.Equal(t, 100, cnt)
+}
+
+func TestWorldRemoveEntities(t *testing.T) {
+	world := NewWorld(NewConfig().WithCapacityIncrement(16))
+	world.SetListener(func(e EntityEvent) {})
+
+	posID := ComponentID[position](&world)
+	rotID := ComponentID[rotation](&world)
+
+	query := world.newEntities(100, posID)
+	assert.Equal(t, 100, query.Count())
+	query.Close()
+
+	query = world.newEntities(100, posID, rotID)
+	assert.Equal(t, 100, query.Count())
+	query.Close()
+
+	query = world.Query(All())
+	assert.Equal(t, 200, query.Count())
+	query.Close()
+
+	filter := All(posID).Exact()
+	world.removeEntities(&filter)
+
+	query = world.Query(All())
+	assert.Equal(t, 100, query.Count())
+	query.Close()
+
+	query = world.Query(All(posID, rotID))
+	assert.Equal(t, 100, query.Count())
+	query.Close()
+}
+
 func TestWorldLock(t *testing.T) {
 	world := NewWorld()
 
@@ -776,7 +917,7 @@ func BenchmarkGetResourceShortcut(b *testing.B) {
 	_ = res
 }
 
-func BenchmarkAdd_10_000_New(b *testing.B) {
+func BenchmarkNewEntities_10_000_New(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		world := NewWorld(NewConfig().WithCapacityIncrement(1024))
 
@@ -789,7 +930,19 @@ func BenchmarkAdd_10_000_New(b *testing.B) {
 	}
 }
 
-func BenchmarkAdd_10_000_Reset(b *testing.B) {
+func BenchmarkNewEntitiesBatch_10_000_New(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		world := NewWorld(NewConfig().WithCapacityIncrement(1024))
+
+		posID := ComponentID[position](&world)
+		velID := ComponentID[velocity](&world)
+
+		q := world.newEntities(10000, posID, velID)
+		q.Close()
+	}
+}
+
+func BenchmarkNewEntities_10_000_Reset(b *testing.B) {
 	b.StopTimer()
 	world := NewWorld(NewConfig().WithCapacityIncrement(1024))
 
@@ -806,5 +959,64 @@ func BenchmarkAdd_10_000_Reset(b *testing.B) {
 		for i := 0; i < 10000; i++ {
 			_ = world.NewEntity(posID, velID)
 		}
+	}
+}
+
+func BenchmarkNewEntitiesBatch_10_000_Reset(b *testing.B) {
+	b.StopTimer()
+	world := NewWorld(NewConfig().WithCapacityIncrement(1024))
+
+	posID := ComponentID[position](&world)
+	velID := ComponentID[velocity](&world)
+
+	for i := 0; i < 10000; i++ {
+		_ = world.NewEntity(posID, velID)
+	}
+	b.StartTimer()
+
+	for i := 0; i < b.N; i++ {
+		world.Reset()
+		q := world.newEntities(10000, posID, velID)
+		q.Close()
+	}
+}
+
+func BenchmarkRemoveEntities_10_000(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		world := NewWorld(NewConfig().WithCapacityIncrement(10000))
+
+		posID := ComponentID[position](&world)
+		velID := ComponentID[velocity](&world)
+
+		entities := make([]Entity, 10000)
+		q := world.newEntities(10000, posID, velID)
+
+		cnt := 0
+		for q.Next() {
+			entities[cnt] = q.Entity()
+			cnt++
+		}
+
+		b.StartTimer()
+
+		for _, e := range entities {
+			world.RemoveEntity(e)
+		}
+	}
+}
+
+func BenchmarkRemoveEntitiesBatch_10_000(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		world := NewWorld(NewConfig().WithCapacityIncrement(10000))
+
+		posID := ComponentID[position](&world)
+		velID := ComponentID[velocity](&world)
+
+		q := world.newEntities(10000, posID, velID)
+		q.Close()
+		b.StartTimer()
+		world.removeEntities(All(posID, velID))
 	}
 }